requirements.txt
streamlit
tensorflow
tensorflow-hub
numpy
pillow
scipy